python/requirements.txt
azure-core==1.4.0
azure-eventhub==5.0.1
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
python-dotenv==0.13.0
requests==2.23.0
six==1.14.0
uamqp==1.2.6
urllib3==1.25.9
